requirements.txt
pytube>=15.0.0
youtube-dl>=2021.12.17
acrcloud-sdk-python>=1.3.0
audd>=0.1.0
shazamio>=0.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
spotipy>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth>=2.22.0
google-api-python-client>=2.108.0 